pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "egile-agent-prospectfinder"
version = "0.1.0"
description = "ProspectFinder plugin for Egile Agent Core - integrates with egile-mcp-prospectfinder MCP server"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["agent", "plugin", "mcp", "prospect-finder", "business-intelligence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-mcp-prospectfinder @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-mcp-prospectfinder",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
all = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-mcp-prospectfinder @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-mcp-prospectfinder",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/jpoullet2000/egile-agent-prospectfinder"
Repository = "https://github.com/jpoullet2000/egile-agent-prospectfinder"
Issues = "https://github.com/jpoullet2000/egile-agent-prospectfinder/issues"

# Entry point for plugin discovery
[project.entry-points."egile_agent_core.plugins"]
prospectfinder = "egile_agent_prospectfinder:ProspectFinderPlugin"

# Console scripts for easy launching
[project.scripts]
prospectfinder = "egile_agent_prospectfinder:run_all"
prospectfinder-mcp = "egile_agent_prospectfinder.run_mcp:run_mcp_only"
prospectfinder-agent = "egile_agent_prospectfinder.run_agent:run_agent_only"

[tool.hatch.build.targets.wheel]
packages = ["src/egile_agent_prospectfinder"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
